add editing ability and filters

+++ ToDo/ToDoList/Core/Storage/StorageManager.swift
@@ -15,7 +15,10 @@ protocol StorageManagerProtocol {
     
     func createTracker(with tracker: TaskModel)
     func getAllTrackers() -> [TaskModel]
+    func getTasksForToday() -> [TaskModel]
     func deleteTracker(by id: UUID)
+    func updateTrackerCompletion(by id: UUID)
+    func updateTrackerData(withData newData: TaskModel)
 }
 
 final class StorageManager: StorageManagerProtocol {
@@ -27,7 +30,7 @@ final class StorageManager: StorageManagerProtocol {
     
     // MARK: - Core Data stack
     private lazy var persistentContainer: NSPersistentContainer = {
-        let container = NSPersistentContainer(name: "Tracker")
+        let container = NSPersistentContainer(name: "Storage")
         container.loadPersistentStores(completionHandler: { (_,
                                                              error) in
             if let error = error as NSError? {
@@ -54,6 +57,7 @@ final class StorageManager: StorageManagerProtocol {
         let trackerEntity = TrackerCoreData(context: context)
         trackerEntity.id = tracker.id
         trackerEntity.toDo = tracker.todo
+        trackerEntity.completed = tracker.completed
         trackerEntity.comment = tracker.comment
         trackerEntity.startDate = tracker.startDate
         trackerEntity.endDate = tracker.endDate
@@ -91,20 +95,75 @@ final class StorageManager: StorageManagerProtocol {
         }
     }
     
+    func getTasksForToday() -> [TaskModel] {
+        let calendar = Calendar.current
+        let todayStart = calendar.startOfDay(for: Date())
+        let todayEnd = calendar.date(byAdding: .day, value: 1, to: todayStart)!
+        
+        let fetchRequest: NSFetchRequest<TrackerCoreData> = TrackerCoreData.fetchRequest()
+        fetchRequest.predicate = NSPredicate(format: "startDate >= %@ AND startDate < %@",
+                                             todayStart as NSDate,
+                                             todayEnd as NSDate)
+        do {
+            let trackerEntities = try context.fetch(fetchRequest)
+            let tasks = trackerEntities.map { TaskModel(from: $0) }
+            storageRelay.accept(tasks)
+            return trackerEntities.map { TaskModel(from: $0) }
+        } catch {
+            return []
+        }
+    }
+    
     func deleteTracker(by id: UUID) {
             let fetchRequest: NSFetchRequest<TrackerCoreData> = TrackerCoreData.fetchRequest()
             fetchRequest.predicate = NSPredicate(format: "id == %@",
                                                  id as CVarArg)
-            
             do {
                 let trackerEntities = try context.fetch(fetchRequest)
-                for trackerEntity in trackerEntities {
+                if let trackerEntity = trackerEntities.first {
                     context.delete(trackerEntity)
                 }
                 saveContext()
-                storageRelay.accept(self.getAllTrackers())
+                self.storageRelay.accept(self.getAllTrackers())
             } catch {
                 fatalError("Failed to delete task: \(error)")
             }
         }
+    
+    func updateTrackerCompletion(by id: UUID) {
+        let fetchRequest: NSFetchRequest<TrackerCoreData> = TrackerCoreData.fetchRequest()
+        fetchRequest.predicate = NSPredicate(format: "id == %@",
+                                             id as CVarArg)
+        do {
+            let trackerEntities = try context.fetch(fetchRequest)
+            if let trackerEntity = trackerEntities.first {
+                trackerEntity.completed = !trackerEntity.completed
+            }
+            saveContext()
+            self.storageRelay.accept(self.getAllTrackers())
+        } catch {
+            fatalError("Failed to update tracker completion: \(error)")
+        }
+    }
+    
+    func updateTrackerData(withData newData: TaskModel) {
+        let id = newData.id
+        let fetchRequest: NSFetchRequest<TrackerCoreData> = TrackerCoreData.fetchRequest()
+        fetchRequest.predicate = NSPredicate(format: "id == %@",
+                                             id as CVarArg)
+        
+        do {
+            let trackerEntities = try context.fetch(fetchRequest)
+            if let trackerEntity = trackerEntities.first {
+                trackerEntity.toDo = newData.todo
+                trackerEntity.comment = newData.comment
+                trackerEntity.startDate = newData.startDate
+                trackerEntity.endDate = newData.endDate
+            }
+            saveContext()
+            storageRelay.accept(self.getAllTrackers())
+        } catch {
+            fatalError("Failed to update tracker completion: \(error)")
+        }
+    }
 }
